2016-01-15 11:14 UTC+0100 Viktor Szakats (vszakats users.noreply.github.com)   * contrib/hbssl/pem.c     * cast
  * src/rdd/usrrdd/rdds/arrayrdd.prg
    * use preferred array index format in remaining cases

  * src/rdd/usrrdd/rdds/hscdx.prg
    * casing of public functions

  + contrib/rdduser/rdduser.hbc
  + contrib/rdduser/rdduser.hbp
  + contrib/rdduser/rdduser.hbx
    + add new contrib 'rdduser'. It offers the exact same content
      as 'hbuddusr' library in core, but having it in the contrib
      area offer much more flexibility and visibility.
      Old library is kept while testing this.

+++ contrib/hbssl/pem.c
@@ -122,7 +122,7 @@ static void hb_PEM_read_bio( PEM_READ_BIO * func )
 
       if( pBuffer )
       {
-         OPENSSL_cleanse( pBuffer, nSize );
+         OPENSSL_cleanse( pBuffer, ( size_t ) nSize );
          hb_xfree( pBuffer );
       }
    }
